test: split the diesel tests into their own integration test
Keeps postgis.rs focused on the DB-backed postgres and sqlx tests, and
lets the diesel tests declare narrower required-features so they run
under --features with-postgis-diesel alone rather than inheriting the
with-geo requirement.

The postgis_diesel_macros compile check stays in postgis.rs next to the
postgres, sqlx and gpkg ones, since it uses the PointZ fixture defined
there.
